requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-extensions==4.1
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
gunicorn==23.0.0
idna==3.11
logging==0.4.9.6
packaging==25.0
psycopg2==2.9.11
python-dotenv==1.2.1
python-json-logger==4.0.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
